requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
Requests==2.32.3
streamlit==1.35.0
google.generativeai==0.7.1
pypdf==4.2.0
openpyxl==3.1.5
lxml
streamlit-keyup
duckdb